+++ .prettierrc
@@ -1,4 +1,6 @@
 {
+  "bracketSpacing": true,
+  "tabWidth": 4,
   "semi": true,
   "trailingComma": "all",
   "singleQuote": true
