+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "rimraf ./dist/ && node build.js",
     "test-pr": "node tests/pull_request.js",
-    "test": "node tests/all.js"
+    "test": "node tests/test_all.js"
   },
   "repository": {
     "type": "git",
