+++ tests/common.js
